requirements.txt
# Core LLM stack
git+https://github.com/huggingface/transformers.git
git+https://github.com/TimDettmers/bitsandbytes.git

accelerate==0.28.0
huggingface-hub>=0.30.0
sentence-transformers==2.7.0

# LangChain ecosystem
langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.33
langserve[all]  # pulls compatible fastapi & deps
langchain-chroma==0.1.1
chromadb==0.4.24
langsmith>=0.1.0

# Vector stores & RAG
faiss-cpu==1.8.0  # WARNING: no Py3.12 prebuilt wheel; may require compiling
pypdf==4.2.0
rapidocr-onnxruntime==1.3.25
unstructured==0.16.11

# API / Middleware
fastapi==0.110.2
uvicorn==0.29.0
beautifulsoup4==4.12.3

# Fixes for known compatibility issues
httpx<0.26  # required for langserve (VerifyTypes error fix)
typing-extensions>=4.5.0